Remove standards violating GPG binary location caching
Fixes #1662

RELEASE_NOTES=[BUGFIX] Remove GPG location caching
