+++ internal/backend/crypto/gpg/cli/binary_others.go
@@ -1,0 +1,14 @@
+// +build !windows
+
+package cli
+
+import (
+	"os/exec"
+)
+
+func detectBinary(name string) (string, error) {
+	if name == "" {
+		name = "gpg"
+	}
+	return exec.LookPath(name)
+}
