--- internal/backend/crypto/gpg/cli/gpg_others.go
@@ -1,11 +1,0 @@
-// +build !windows
-
-package cli
-
-func detectBinaryCandidates(bin string) ([]string, error) {
-	bins := []string{"gpg2", "gpg1", "gpg"}
-	if bin != "" {
-		bins = append(bins, bin)
-	}
-	return bins, nil
-}
